code/gittyup/pyproject.toml
[project]
name = "gittyup"
version = "0.1.0"
description = "A professional-grade CLI tool that automatically discovers and updates all Git repositories within a directory tree"
readme = "readme.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Michael Kennedy"}
]
keywords = ["git", "cli", "automation", "repository", "update"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Version Control :: Git",
]

dependencies = [
    "colorama>=0.4.6",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.6.0",
]

[project.scripts]
gittyup = "gittyup.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gittyup"]

